% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunction.R
\name{absoluteinner}
\alias{absoluteinner}
\title{create sub folders in the absolute path folder}
\usage{
absoluteinner(x, var)
}
\arguments{
\item{x}{is the absolute path set in .abspath}

\item{var}{is the name of the folder with specific variables.}
}
\value{
sub folder in the absolute path
}
\description{
create sub folders in the absolute path folder
}
