% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfunction.R
\name{before_u_start}
\alias{before_u_start}
\title{Steps to follow in using the Freshwater Information Platform}
\usage{
before_u_start()
}
\value{
\code{list} steps to follow in using the FIP
}
\description{
Steps to follow in using the Freshwater Information Platform
}
\examples{

#b4us <- before_u_start()

}
\author{
Anthony Basooma
}
