% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_getfiles.R
\name{collatedata}
\alias{collatedata}
\title{Data download from the Freshwater Information Platform.}
\usage{
collatedata(
  taxa,
  ecotraits = NULL,
  taxaorder = NULL,
  token = NULL,
  warn = FALSE,
  multiple = FALSE,
  parallel = FALSE,
  cores = 2,
  quietly = FALSE
)
}
\arguments{
\item{taxa}{\code{string}. The taxa group to download from the platform.
The allowed group include \code{"fi", "mi", "pp", "pb", "di","mp"} for fishes, macroinvertebrates, phytoplankton,
phytobenthos, diatoms, and macrophytes. Multiple groups allowed such as \code{'pp', 'di'}.
\itemize{
\item{\code{pp: }}{Pytoplankton.}
\item{\code{mp: }}{Macrophtytes}
\item{\code{mi: }}{Macroinvertebrates}
\item{\code{fi: }}{Fishes}
\item{\code{di: }}{Diatoms}
\item{\code{pb: }}{Phytobenthos without diatoms}
}}

\item{ecotraits}{\code{vector}. Selected traits that should be downloaded for particular taxa group. Check \code{\link{fw_ecoparamdb}} for the allowed
traits in the database.}

\item{taxaorder}{\code{vector}. If \code{taxa} is \code{mi}, the \code{taxaorder} must be indicated for data to be downloaded.
The different macroinvertebrates orders allowed can be obtained using \code{\link{fw_orders}} function.}

\item{token}{\code{string}. This is a required parameter to allow user authentication with the platform. To get the token, use
\code{\link{before_u_start}} function to get the required steps. Remember that the token is saved in memory such that
the data downloaded is not re-downloaded in the next session.}

\item{warn}{\code{logical} To show species name warning checks and traits cleaning. Default \code{FALSE}.}

\item{multiple}{\code{logical}. Either \code{TRUE} if multiple taxa groups are considered. Default is \code{FALSE} for only
single taxon group.}

\item{parallel}{\code{logcial}. Either \code{TRUE} to allow parallelisation most especially if multiple groups are considered.
If multiple is set to TRUE and parallel to FALSE, a slow mode of data download will be executed.}

\item{cores}{\code{interger}. The number computer cores to be used to run while data download and its necessary only when
parallel is set to \code{TRUE}. Default is 2.}

\item{quietly}{\code{logical}. If \code{FALSE}, a message showing the slow mode of data collation
will be presented to the user. Default \code{TRUE}.}
}
\value{
\code{list}. Lists of taxa groups data downloaded from the Freshwater Information Platform.
}
\description{
Data download from the Freshwater Information Platform.
}
\details{
The \code{getdata} is used as a wrapper to enable the user to download data from multiple groups
and allowing parallelising the \code{\link{getfiles}} function. The user can change the number of cores to improve the speed of
data downloaded. All data downloaded in this function by the user  is cached in the \code{taxongroups} folder, which
is automatically created in the working directory when loading the package. The data is permanently cached, even when
the session is restarted. The user can use the \code{\link{fw_decache}} function to remove the files. However, to extract the data
please check both \code{\link{extract_traits}} and \code{\link{fw_ecoparameters}}.
}
\examples{

\dontrun{

ppdata <- 1
}

}
