% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_dbchecks.R
\name{compare_traits}
\alias{compare_traits}
\title{Traits comparison for the user and standard triat names.}
\usage{
compare_traits(traits)
}
\arguments{
\item{traits}{\code{string}. Trait name provided by the user that will be compared with the
standard names in the database.}
}
\value{
\code{string} accepted trait names based on the database
}
\description{
Traits comparison for the user and standard triat names.
}
