% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoparameters}
\alias{ecoparameters}
\title{Standard date formats}
\format{
A \code{list} Species ecoparameters.
}
\usage{
data(ecoparameters)
}
\description{
A \code{tibble} Dates standard dataframe. Extensible by the package author
}
\details{
Standard architecture of the API
}
\examples{
\dontrun{

data("ecoparameters")
ecoparameters
}

}
\author{
Anthony Basooma (anthony.basooma@boku.ac.at)
}
\keyword{datasets}
