% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_extractdata.R
\name{extract_traits}
\alias{extract_traits}
\title{Extracting the traits from the downloaded data.}
\usage{
extract_traits(
  data,
  taxa,
  ecotraits = NULL,
  taxaorder = NULL,
  token = NULL,
  multiple = FALSE,
  parallel = FALSE,
  cores = NULL,
  quietly = FALSE,
  pct = 80,
  errorness = 20,
  warn = FALSE
)
}
\arguments{
\item{data}{\code{vector}. The list or vector with species names for which ecological references needs to be extracted from the
database.}

\item{taxa}{\code{string}. The taxa group to download from the platform.
The allowed group include \code{"fi", "mi", "pp", "pb", "di","mp"} for fishes, macroinvertebrates, phytoplankton,
phytobenthos, diatoms, and macrophytes. Multiple groups allowed such as \code{'pp', 'di'}.
\itemize{
\item{\code{pp: }}{Pytoplankton.}
\item{\code{mp: }}{Macrophtytes}
\item{\code{mi: }}{Macroinvertebrates}
\item{\code{fi: }}{Fishes}
\item{\code{di: }}{Diatoms}
\item{\code{pb: }}{Phytobenthos without diatoms}
}}

\item{ecotraits}{\code{vector}. Selected traits that should be downloaded for particular taxa group. Check \code{\link{fw_ecoparamdb}} for the allowed
traits in the database.}

\item{taxaorder}{\code{vector}. If \code{taxa} is \code{mi}, the \code{taxaorder} must be indicated for data to be downloaded.
The different macroinvertebrates orders allowed can be obtained using \code{\link{fw_orders}} function.}

\item{token}{\code{string}. This is a required parameter to allow user authentication with the platform. To get the token, use
\code{\link{before_u_start}} function to get the required steps. Remember that the token is saved in memory such that
the data downloaded is not re-downloaded in the next session.}

\item{multiple}{\code{logical}. Either \code{TRUE} if multiple taxa groups are considered. Default is \code{FALSE} for only
single taxon group.}

\item{parallel}{\code{logcial}. Either \code{TRUE} to allow parallelisation most especially if multiple groups are considered.
If multiple is set to TRUE and parallel to FALSE, a slow mode of data download will be executed.}

\item{cores}{\code{interger}. The number computer cores to be used to run while data download and its necessary only when
parallel is set to \code{TRUE}. Default is 2.}

\item{quietly}{\code{logical}. If \code{FALSE}, a message showing the slow mode of data collation
will be presented to the user. Default \code{TRUE}.}

\item{pct}{\code{numeric}. The number used as a cutoff to infer similarity of the user provided name and what is found in the database.
The higher the percentage, the higher the similarity the species name provided by the user and the one in the database. \code{pct}
ranges from 0 to 100 but the default is 80 to ensure that wrong names are not selected at low similarity percentage cutoff.}

\item{errorness}{\code{numeric} Similar to \code{pct}, \code{errorness} parameter uses the distance differences between the user-provided
names and all the taxa group species standard names. The lower the percentage error, the higher the similarity in the species
names provided. Default is 20 and beyond 30, a warning is showed to avoid wrong species replace the user provided name, which leads
to extracting wrong traits.}

\item{warn}{\code{logical} To show species name warning checks and traits cleaning. Default \code{FALSE}.}
}
\value{
\code{dataframe} A dataframe species traits for all orders.
}
\description{
Extracting the traits from the downloaded data.
}
