% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{fip_traits}
\alias{fip_traits}
\title{Title}
\usage{
fip_traits(
  data,
  spcol = NULL,
  group = "fi",
  traits = "all",
  taxonomy = NULL,
  taxaorder = NULL,
  token = NULL,
  multiple = FALSE,
  parallel = FALSE,
  cores = 2,
  quietly = FALSE,
  level = NULL,
  wide = FALSE,
  select = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{grouptaxa}{}
}
\description{
Title
}
