% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fuzzy_codes}
\alias{fuzzy_codes}
\title{Internal package and create dummy codes}
\usage{
fuzzy_codes(trait, FD)
}
\arguments{
\item{trait}{\code{dataframe} Data matrix}

\item{FD}{\code{logical} Either to compute functional diversity indices or not.}
}
\description{
Internal package and create dummy codes
}
