% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fw_be4ustart}
\alias{fw_be4ustart}
\title{Steps to follow in using the Freshwater Information Platform}
\usage{
fw_be4ustart()
}
\value{
\code{list} steps to follow in using the FWDB
}
\description{
Steps to follow in using the Freshwater Information Platform
}
\examples{

#b4us <- fw_be4ustart()

}
\author{
Anthony Basooma
}
