% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fw_cite}
\alias{fw_cite}
\title{Get database citations}
\usage{
fw_cite(cachefolder = "cache")
}
\arguments{
\item{cachefolder}{\code{string}. The root path were the cached data will be saved on the user PC.
If the path is not provided, the cached information will be saved in the current
working directly.}
}
\value{
string of the package citation
}
\description{
Get database citations
}
