% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_dbguide.R
\name{fw_dbguide}
\alias{fw_dbguide}
\title{Standard table with taxonomic groups and their traits explanations.}
\usage{
fw_dbguide(organismgroup = NULL, traits = NULL)
}
\arguments{
\item{organismgroup}{\code{string or vector}. Taxa group names to aid the users in filtering the standard
table for species traits and their explanations.}

\item{traits}{\code{string or vector}. Traits that can be filtered from the databases, given the organism group
indicated in \code{organismgroup} parameter.}
}
\value{
\code{dataframe} A dataset with taxonomic groups, traits and their explanations.
}
\description{
Standard table with taxonomic groups and their traits explanations.
}
\examples{

\dontrun{

dbase <- fw_dbguide()

}

}
