% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_decache.R
\name{fw_decache}
\alias{fw_decache}
\title{Decache tokens and data files.}
\usage{
fw_decache(type = "token", action = "reset")
}
\arguments{
\item{type}{\code{string}. Either \code{taxa} or \code{token} to be removed from the data.}

\item{action}{\code{string}. Either \code{reset} to remove token or data files or
\code{destroy} to remove all the files and cache directory.}
}
\value{
\code{message}. files or token cache token or taxa files and directory removed.
}
\description{
Decache tokens and data files.
}
