% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_ecoparamdb.R
\name{fw_ecoparamdb}
\alias{fw_ecoparamdb}
\title{Standard table with taxonomic groups and their traits explanations.}
\usage{
fw_ecoparamdb(taxagroup = NULL, traits = NULL)
}
\arguments{
\item{taxagroup}{\code{string or vector}. Taxa group names to aid the users in filtering the standard
table for species traits and their explanations.}

\item{traits}{\code{string or vector}. Traits that can be filtered from the databases, given the taxonomic group
indicated in \code{taxagroup} parameter.}
}
\value{
\code{dataframe} A dataset with taxonomic groups, traits and their explanations.
}
\description{
Standard table with taxonomic groups and their traits explanations.
}
