% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_ecoparameters.R
\name{fw_ecoparameters}
\alias{fw_ecoparameters}
\title{Arranging and user friendly selection of traits at different taxanomic levels.}
\usage{
fw_ecoparameters(
  data,
  taxa,
  token,
  spcol = NULL,
  taxacol = NULL,
  ecotraits = NULL,
  traits = "all",
  level = NULL,
  taxaorder = NULL,
  multiple = FALSE,
  parallel = FALSE,
  cores = 2,
  quietly = FALSE,
  wide = FALSE,
  selectvalue = NULL,
  descvalue = NULL,
  na.rm = TRUE,
  merge = FALSE,
  warn = FALSE,
  errorness = 20,
  pct = 80
)
}
\arguments{
\item{data}{\code{vector}. The list or vector with species names for which ecological references needs to be extracted from the
database.}

\item{taxa}{\code{string}. The taxa group to download from the platform.
The allowed group include \code{"fi", "mi", "pp", "pb", "di","mp"} for fishes, macroinvertebrates, phytoplankton,
phytobenthos, diatoms, and macrophytes. Multiple groups allowed such as \code{'pp', 'di'}.
\itemize{
\item{\code{pp: }}{Pytoplankton.}
\item{\code{mp: }}{Macrophtytes}
\item{\code{mi: }}{Macroinvertebrates}
\item{\code{fi: }}{Fishes}
\item{\code{di: }}{Diatoms}
\item{\code{pb: }}{Phytobenthos without diatoms}
}}

\item{token}{\code{string}. This is a required parameter to allow user authentication with the platform. To get the token, use
\code{before_u_start()} function to get the required steps. Remember that the token is saved in memory such that
the data downloaded is not re-downloaded in the next session.}

\item{spcol}{\code{string}. If the data is a dataframe, the species column is required and provided in this parameter.
The column should have complete species name and not genus and species provided separately.}

\item{taxacol}{\code{string} If the data is a dataframe, and more than one taxonomic group exists in the data, the
the \code{taxacol} is required to iterate over the taxonomic groups separately.}

\item{ecotraits}{\code{vector}. Selected traits that should be downloaded for particular taxa group. Check \code{\link{fw_ecoparamdb}} for the allowed
traits in the database.}

\item{traits}{\code{string or vector}. If \code{all} is indicated, then all the traits will be extracted. Otherwise,
individual traits can be indicated in a vector format. Check the allowed traits in \code{\link{fw_ecoparamdb}} function and identify
all the traits allowed for each group and their explanation.}

\item{level}{\code{string}. The taxonomic orders allowed for each species including \code{species, genus, order or family}.}

\item{taxaorder}{\code{vector}. If \code{taxa} is \code{mi}, the \code{taxaorder} must be indicated for data to be downloaded.
The different macroinvertebrates orders allowed can be obtained using \code{fw_orders()} function.}

\item{multiple}{\code{logical}. Either \code{TRUE} if multiple taxa groups are considered. Default is \code{FALSE} for only
single taxon group.}

\item{parallel}{\code{logcial}. Either \code{TRUE} to allow parallelisation most especially if multiple groups are considered.
If multiple is set to TRUE and parallel to FALSE, a slow mode of data download will be executed.}

\item{cores}{\code{interger}. The number computer cores to be used to run while data download and its necessary only when
parallel is set to \code{TRUE}. Default is 2.}

\item{quietly}{\code{logical}. If \code{FALSE}, a message showing the slow mode of data collation
will be presented to the user. Default \code{TRUE}.}

\item{wide}{\code{logical}. If \code{TRUE}, then the output is spread to a wider or spread format for each unique species and
taxonomic groups.}

\item{selectvalue}{\code{vector}. To allow user selection within the traits, for example, for fishes if catchment region is considered
then the species native in the Danube region can be selected and retained.}

\item{descvalue}{\code{vector}. To allow traits selection within the traits, for example, for fishes and catchment regions is
is considered, 'Danu' can be selected.}

\item{na.rm}{\code{logical} If \code{TRUE}, then the traits with no data will be removed from the output dataset.
Default \code{TRUE}.}

\item{merge}{\code{logical}. If the data is a dataframe and not list or vector, merge allows to
affix the species ecological parameters on the user dataframe. Default \code{FALSE} to return
only the species data but not the whole user input dataset.}

\item{warn}{\code{logical} To show species name warning checks and traits cleaning. Default \code{FALSE}.}

\item{errorness}{\code{numeric} Similar to \code{pct}, \code{errorness} parameter uses the distance differences between the user-provided
names and all the taxa group species standard names. The lower the percentage error, the higher the similarity in the species
names provided. Default is 20 and beyond 30, a warning is showed to avoid wrong species replace the user provided name, which leads
to extracting wrong traits.}

\item{pct}{\code{numeric}. The number used as a cutoff to infer similarity of the user provided name and what is found in the database.
The higher the percentage, the higher the similarity the species name provided by the user and the one in the database. \code{pct}
ranges from 0 to 100 but the default is 80 to ensure that wrong names are not selected at low similarity percentage cutoff.}
}
\value{
\code{dataframe} A dataframe species traits at a selected taxonomic level.
}
\description{
Arranging and user friendly selection of traits at different taxanomic levels.
}
\author{
{
Anthony Basooma \email{anthony.basooma@boku.ac.at}
}
}
