% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_fdendro.R
\name{fw_fdendro}
\alias{fw_fdendro}
\title{For computing the functional dendogram}
\usage{
fw_fdendro(fwdata, method = "average", k = 4, plot = FALSE)
}
\arguments{
\item{fwdata}{\code{list} A list from the \code{fw_fetchdata} function is the only data type accepted.}

\item{method}{\code{string} Clustering method, including ward, average, and complete}

\item{k}{\code{integer} Determine the number of cluster the user wants to output after data clustering. The default is 4.}

\item{plot}{\code{logical} Either \code{TRUE} to show the plot of functional dendogram. Defualt is \code{FALSE}.}
}
\value{
dataframe, plot
}
\description{
For computing the functional dendogram
}
\examples{

\dontrun{
  fishtraits <- fw_fetchdata(data = speciesdata,
                        ecoparams = c('rheophily habitat', 'spawning habitat',
                        'feeding diet adult'),
                        taxonomic_column = 'scientificName',
                        organismgroup = 'fi')


 head(fdendoclust, 3)

 table(fdendoclust$cluster)

}
}
