% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_fetchdata.R
\name{fw_fetchdata}
\alias{fw_fetchdata}
\title{Arranging and user friendly selection of traits at different taxonomic levels.}
\usage{
fw_fetchdata(
  data,
  organismgroup,
  token,
  spcol = NULL,
  groupcol = NULL,
  ecoparams = NULL,
  lat = NULL,
  lon = NULL,
  cachepath = NULL,
  inform = FALSE,
  wide = FALSE,
  na.rm = TRUE,
  merge = FALSE,
  warn = FALSE,
  sanitize = FALSE,
  errorness = 27,
  pct = 80
)
}
\arguments{
\item{data}{\code{vector}. The list or vector with species names for which ecological references needs to be extracted from the
database.}

\item{organismgroup}{\code{string}. The organism group to download from the platform.
The allowed group include \code{"fi", "mi", "pp", "pb", "di","mp"} for fishes,
macroinvertebrates, phytoplankton,
phytobenthos, diatoms, and macrophytes, respectively. Multiple groups allowed such as \code{'pp', 'di'}.
\itemize{
\item{\code{pp}: Pytoplankton.}
\item{\code{mp}: Macrophtytes}
\item{\code{mi}: Macroinvertebrates}
\item{\code{fi}: Fishes}
\item{\code{di}: Diatoms}
\item{\code{pb}: Phytobenthos without diatoms}
}}

\item{token}{\code{string}. This is a required parameter to allow user authentication with the platform. To get the token, use
\code{\link{fw_be4ustart}} function to get the required steps. Remember that the token is saved in memory such that
the data downloaded is not re-downloaded in the next session.}

\item{spcol}{\code{string}. If the data is a dataframe, the species column is required and provided in this parameter.
The column should have complete species name and not genus and species provided separately.}

\item{groupcol}{\code{string} If the data is a dataframe, and more than one taxonomic group exists in the data, the
the \code{groupcol} is required to iterate over the taxonomic groups separately.}

\item{ecoparams}{\code{vector}. Selected traits that should be downloaded for particular organismgroup group. Check \code{\link{fw_dbguide}} for the allowed
traits in the database.}

\item{lat, lon}{\code{string}. These are columns that have both the longitude and latitude
values in the species dataset.}

\item{cachepath}{\code{string}. The root path were the cached data will be saved on the user PC.
If the path is not provided, the cached information will be saved in the current
working directly.}

\item{inform}{\code{logical}. To indicate if the token is successfully generated. Default \code{TRUE}.}

\item{wide}{\code{logical}. If \code{TRUE}, then the output is spread to a wider or spread format for each unique species and
taxonomic groups.}

\item{na.rm}{\code{logical} If \code{TRUE}, then the traits with no data will be removed from the output dataset.
Default \code{TRUE}.}

\item{merge}{\code{logical}. If the data is a dataframe and not list or vector, merge allows to
affix the species ecological parameters on the user dataframe. Default \code{FALSE} to return
only the species data but not the whole user input dataset.}

\item{warn}{\code{logical} To show species name warning checks and traits cleaning. Default \code{FALSE}.}

\item{sanitize}{\code{logical} Either \code{TRUE} to return complete trait names but not abbreviations which is normally returned
from the database.}

\item{errorness}{\code{numeric} Similar to \code{pct}, \code{errorness} parameter uses the distance differences between the user-provided
names and all the taxa group species standard names. The lower the percentage error, the higher the similarity in the species
names provided. Default is 20 and beyond 30, a warning is showed to avoid wrong species replace the user provided name, which leads
to extracting wrong traits.}

\item{pct}{\code{numeric}. The number used as a cutoff to infer similarity of the user provided name and what is found in the database.
The higher the percentage, the higher the similarity the species name provided by the user and the one in the database. \code{pct}
ranges from 0 to 100 but the default is 80 to ensure that wrong names are not selected at low similarity percentage cutoff.}
}
\value{
\code{dataframe} A dataframe species traits at a selected taxonomic level.
}
\description{
Arranging and user friendly selection of traits at different taxonomic levels.
}
\examples{

\dontrun{

library(httr2)

sacrambled_api <- "p6-9gAvYXveyC_B_0hTzyYl5FwLRwZEPD-ZE9Y4KrIBstgNc8K2Mr4u6t39LGZ2E1m9SOw"
#scrambled_api_key

#download fish catchment region data

apikeydecrypted <- fw_loadapikey(test = TRUE, sacrambled_apikey = sacrambled_api,
                              fwtraitskey =  'FWTRAITS_KEY')

tokendata <- fw_token(key= apikeydecrypted, seed = 1234)

fishdata <- fw_fetchdata(data = 'Abramis brmaam', organismgroup = 'fi',
                             ecoparams = 'catchment region',
                             token = tokendata)#the species spelling is checked
}

}
\seealso{
\code{\link{fw_token}}, \code{\link{fw_searchdata}}, \code{\link{fw_split}}
}
\author{
{
Anthony Basooma \email{anthony.basooma@boku.ac.at}
}
}
