% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_setapikey.R
\name{fw_keyload}
\alias{fw_keyload}
\title{For loading the API key from the environment.}
\usage{
fw_keyload(keyvar = "API_KEY")
}
\arguments{
\item{keyvar}{\code{string} Is the API name as indicated in the user environment. Default is \code{"API_KEY"}.}
}
\value{
string of API
}
\description{
For loading the API key from the environment.
}
