% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fw_loadapikey}
\alias{fw_loadapikey}
\title{To load the API key}
\usage{
fw_loadapikey(test = FALSE, encrytedkey = NULL, fwtraitskey = NULL)
}
\arguments{
\item{test}{\code{logical}. If TRUE, then function is used with \code{secret_make_key}
to encrypt the API key while publicly testing shared code.}

\item{encrytedkey}{\code{alphanumeric} An alphanumeric string generated from \code{secret_make_key} fucntion
while testing the codes.}

\item{fwtraitskey}{\code{string} A string that indicates key in the edited Rsession.}
}
\value{
encrypted API key
}
\description{
The function allows the user to protect the key in user system environment to avoid
it being included in the analysis codes.
}
\examples{

\dontrun{
enc_api <- "p6-9gAvYXveyC_B_0hTzyYl5FwLRwZEPD-ZE9Y4KrIBstgNc8K2Mr4u6t39LGZ2E1m9SOw"

#the FWTRAITS_KEY is the unlock key saved in my local environment
#check https://httr2.r-lib.org/articles/wrapping-apis.html for more information

apikey <- httr2::secret_decrypt(encrypted = enc_api, key = 'FWTRAITS_KEY')

#download fish catchment region data
#setting the FWTRAITS_KEY

#run this usethis::edit_r_environ()

apikeydecrypted <- fw_loadapikey(test = TRUE, encrytedkey = enc_api,
                              fwtraitskey =  'FWTRAITS_KEY')

}

}
\seealso{
\code{\link{fw_token}}, \code{\link{fw_fetchdata}}
}
