% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_getfiles.R
\name{fw_orders}
\alias{fw_orders}
\title{To get the allowed macroinvertebrates in the Freshwaterecology.info.}
\usage{
fw_orders()
}
\value{
\code{vector}. A vector of allowed macro invertebrates taxonomic groups that can be set in the \code{taxaorder} parameter while
while getting data.
}
\description{
To get the allowed macroinvertebrates in the Freshwaterecology.info.
}
\examples{

x <- fw_orders()

}
