% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fw_paramlist}
\alias{fw_paramlist}
\title{Function to retrieve all the traits in the database.}
\usage{
fw_paramlist(cachefolder = NULL)
}
\value{
\code{list} List of traits for all the taxa groups and orders in the database.
}
\description{
Function to retrieve all the traits in the database.
}
\author{
Anthony Basooma
}
\keyword{internal}
