% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_setapikey.R
\name{fw_setapikey}
\alias{fw_setapikey}
\title{Function and instructions creating the API_KEY variable in the User environment.}
\usage{
fw_setapikey()
}
\description{
Function and instructions creating the API_KEY variable in the User environment.
}
