% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_searchdata.R
\name{fw_taxagroup}
\alias{fw_taxagroup}
\title{To get the allowed macroinvertebrates in the Freshwaterecology.info.}
\usage{
fw_taxagroup()
}
\value{
\code{vector}. A vector of allowed macro invertebrates taxonomic groups that can be set in the \code{taxagroup} parameter while
while getting data.
}
\description{
To get the allowed macroinvertebrates in the Freshwaterecology.info.
}
\examples{

x <- fw_taxagroup()

}
