% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invertbackbone}
\alias{invertbackbone}
\title{Taxonomic data for species in the freshwaterecology.info database}
\format{
A \code{tibble} 10421 rows and 3 columns.

\itemize{
\item Taxgroup: Higher taxonomic grouping for the species, for example, Bivalvia.
\item Family: Taxonomic classification, e.g., CARDIIDAE
\item Taxon: Lower taxonomic grouping for the species.eg., Parvicardium exiguum
}
}
\usage{
data(invertbackbone)
}
\description{
A \code{tibble}
}
\details{
Standard taxonomic backbone for macroinvertebrates from www.freshwaterecology.info.
All macroinvetebrates species are crosschecked with this database improve quality control
of the species names provided by the user.
}
\examples{

\dontrun{

data("invertbackbone")
invertbackbone
}


}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). www.freshwaterecology.info - An online tool that unifies,
standardizes and codifies more than 20,000 European freshwater organisms and their ecological preferences.
Ecological Indicators, 53, 271-282. https://doi.org/10.1016/j.ecolind.2015.02.007.
}
\keyword{Macroinvertebrates}
