% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{orders}
\alias{orders}
\title{To get the allowed macroinvertebrates in the Freshwater Information Platform.}
\usage{
orders()
}
\value{
\code{vector}. A vector of allowed macro invertebrates taxonomic groups that can be set in the \code{taxaorder} parameter while
while getting data.
}
\description{
To get the allowed macroinvertebrates in the Freshwater Information Platform.
}
\examples{

x <- orders()

}
