% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbenthobackbone}
\alias{pbenthobackbone}
\title{Taxonommic data for the phytobentho species in frehwaterecology.info database.}
\format{
A \code{tibble} 1857 rows and 2 columns.
}
\usage{
data(pbenthobackbone)
}
\description{
A \code{tibble}
}
\details{
fhfhdld fnfmf...
}
\examples{

\dontrun{

data("pbenthobackbone")

pbenthobackbone
}




}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). Www.freshwaterecology.info - An online tool that unifies,
standardises and codifies more than 20,000 European freshwater organisms and their ecological preferences.
Ecological Indicators, 53, 271-282. https://doi.org/10.1016/j.ecolind.2015.02.007.
}
\keyword{Phytobentho}
\keyword{data}
