% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbenthodata}
\alias{pbenthodata}
\title{Taxonomic data for the phytobentho species in www.freshwaterecology.info database.}
\format{
A \code{tibble} The dataset has 1857 rows and 2 columns.

\itemize{
\item Taxgroup: Higher taxonomic grouping for the species, for example, Bacillariophyceae.
\item Taxon: Lower taxonomic grouping for the species.eg., Achnanthes acus, Achnanthes brevipes,
Achnanthes brevipes var. brevipes
}
}
\usage{
data(pbenthodata)
}
\description{
A \code{tibble}
}
\details{
This dataset was extracted from the www.freshwaterecology.info database and formed a standardized
reference for the species names for phytobenthos. Therefore, all user-provided species are
checked across this database to identify whether or not they exist.
}
\examples{

\dontrun{

data("pbenthodata")

pbenthodata
}

}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). Www.freshwaterecology.info - An online tool that unifies,
standardises and codifies more than 20,000 European freshwater organisms and their ecological preferences.
Ecological Indicators, 53, 271-282. https://doi.org/10.1016/j.ecolind.2015.02.007.
}
\keyword{Phytobentho}
\keyword{data}
