% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcheck.R
\name{spcheck}
\alias{spcheck}
\title{Title}
\usage{
spcheck(
  taxadb,
  species = NULL,
  check = FALSE,
  quietly = TRUE,
  taxalist = FALSE,
  errror_on_names = TRUE,
  grname = NULL
)
}
\arguments{
\item{errror_on_names}{}
}
\description{
Title
}
