% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{tokenkey}
\alias{tokenkey}
\title{Access and loading the token key}
\usage{
tokenkey(key = loadapikey(), quietly = TRUE)
}
\arguments{
\item{key}{\code{string} The API key which is automatically loaded using the loadapikey() internal function.}

\item{quietly}{\code{logical}. To indicate if the token is successfully generated. Default \code{TRUE}.}
}
\value{
\code{string} token authentication token key
}
\description{
Access and loading the token key
}
