% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractdata.R
\name{extract_traits}
\alias{extract_traits}
\title{Title}
\usage{
extract_traits(
  species,
  group,
  taxaorder = NULL,
  token = NULL,
  multiple = FALSE,
  parallel = FALSE,
  cores = NULL,
  quietly = FALSE,
  pct = 80,
  errorness = 20
)
}
\description{
Title
}
