% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{fip_traits}
\alias{fip_traits}
\title{Title}
\usage{
fip_traits(
  data,
  spcol = NULL,
  group = "fi",
  taxonomy = NULL,
  level = NULL,
  traits = NULL,
  wide = FALSE,
  all = FALSE,
  select = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{grouptaxa}{}
}
\description{
Title
}
