% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_maps.R
\name{fw_maps}
\alias{fw_maps}
\title{ggplot2 visualization.}
\usage{
fw_maps(
  data,
  output,
  lat = NULL,
  lon = NULL,
  basin,
  params = NULL,
  reduce = FALSE,
  group = NULL,
  bgtheme = "bw",
  bgcolor = "grey50",
  epsg = 4326
)
}
\arguments{
\item{data}{The species data either in sf format or with coordinate columns to enable plotting.}

\item{output}{fetchdata output only accepted.}

\item{lat, lon}{\code{string}. These are columns that have both the longitude and latitude
values in the species dataset.}

\item{basin}{\code{string}. Geographical map to be used as a bounding box for drawing the
visualization maps. The species coordinates must within the indicated bounding box
or basin boundary.}

\item{params}{species ecological parameters selected if there are more than one.}

\item{reduce}{\code{logical}. If TRUE, the coordinates are rounded off which reduces the distinct points
to ease ploting. Default FALSE.}

\item{group}{Particular organism group to filter out to allow visualization.}

\item{bgtheme}{\code{string}. The theme to be used for the map based on the ggplot2 package
themes including \code{theme_bw}, \code{theme_classic}.}

\item{bgcolor}{\code{string}. The color which will be used to fill the maps.}

\item{epsg}{Coordinate reference number. Default is 4326.}
}
\value{
ggplot2 display
}
\description{
ggplot2 visualization.
}
