% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_searchdata.R
\name{fw_searchdata}
\alias{fw_searchdata}
\title{To download data from the Freshwaterecology.info database.}
\usage{
fw_searchdata(
  organismgroup,
  taxagroup = NULL,
  ecoparams = NULL,
  token,
  warn = TRUE,
  parallel = FALSE,
  cores = 3
)
}
\arguments{
\item{organismgroup}{\code{string}. The organismgroup group to download from the platform.
The allowed group include \code{"fi", "mi", "pp", "pb", "di","mp"} for fishes, macroinvertebrates, phytoplankton,
phytobenthos, diatoms, and macrophytes. Multiple groups allowed such as \code{'pp', 'di'}.
\itemize{
\item{\code{pp}: Pytoplankton.}
\item{\code{mp}: Macrophtytes}
\item{\code{mi}: Macroinvertebrates}
\item{\code{fi}: Fishes}
\item{\code{di}: Diatoms}
\item{\code{pb}: Phytobenthos without diatoms}
}}

\item{taxagroup}{\code{vector}. If \code{organismgroup} is \code{mi}, the \code{taxagroup} must be indicated for data to be downloaded.
The different macroinvertebrates orders allowed can be obtained using \code{\link{fw_taxagroup}} function.}

\item{ecoparams}{\code{vector}. Selected traits that should be downloaded for particular organismgroup group. Check \code{\link{fw_dbguide}} for the allowed
traits in the database.}

\item{token}{\code{string}. This is a required parameter to allow user authentication with the platform. To get the token, use
\code{\link{fw_be4ustart}} function to get the required steps. Remember that the token is saved in memory such that
the data downloaded is not re-downloaded in the next session.}

\item{warn}{\code{logical} To show species name warning checks and traits cleaning. Default \code{FALSE}.}

\item{parallel}{\code{logical}. If \code{TRUE} then the parallel data download is enabled.}

\item{cores}{\code{integer} An integer indicating the number of cores to be be used in parallelisng the
the data download. Default is 2.}
}
\value{
List of download species traits
}
\description{
The function provides a seamless access and download of species ecological parameters, traits,
or indicators from the Freshwaterecology.info database. The fucntion allows multiple inclusion
of organism groups, which include macroinvertebrates, fishes, phytoplankton, phytobenthos,
macrophytes, and diatoms. Parallelisation can also enabled to allow faster data download from
for the database severs.
}
\examples{

\dontrun{

#encrypted token for my api key

enc_api <- "p6-9gAvYXveyC_B_0hTzyYl5FwLRwZEPD-ZE9Y4KrIBstgNc8K2Mr4u6t39LGZ2E1m9SOw"

apikey <- httr2::secret_decrypt(encrypted = enc_api, key = 'FWTRAITS_KEY')

#download fish catchment region data
#setting the FWTRAITS_KEY

#run this usethis::edit_r_environ()

apikeydecrypted <- fw_loadapikey(test = TRUE, encrytedkey = enc_api,
                              fwtraitskey =  'FWTRAITS_KEY')

tokendata <- fw_token(key= apikeydecrypted, seed = 1234)

dfpull <- fw_searchdata(organismgroup = 'fi', ecoparams = 'migration', token = tokendata)
}
}
