% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_setapikey.R
\name{fw_setapikey}
\alias{fw_setapikey}
\title{Function and instructions creating the API_KEY variable in the User environment.}
\usage{
fw_setapikey(verbose = TRUE)
}
\arguments{
\item{verbose}{\code{logical} If logical, then the instruction message will show. Default \code{TRUE}.}
}
\value{
string of instructions for saving or setting the API in the R session.
}
\description{
Function and instructions creating the API_KEY variable in the User environment.
}
