% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_token.R
\name{fw_token}
\alias{fw_token}
\title{Access and loading the token key}
\usage{
fw_token(key = loadapikey(), quietly = TRUE, seed = NULL)
}
\arguments{
\item{key}{\code{string} The API key which is automatically loaded using the loadapikey() internal function.}

\item{quietly}{\code{logical}. To indicate if the token is successfully generated. Default \code{TRUE}.}

\item{seed}{\code{integer}. An integer to help track the caching of the access token generated during data collation.
If a user wants to get a new token, then the seed should be changed.}
}
\value{
\code{string} token authentication token key
}
\description{
Access and loading the token key
}
