% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_token.R
\name{fw_token}
\alias{fw_token}
\title{Access and loading the token key}
\usage{
fw_token(key = fw_loadapikey(), quietly = TRUE, seed = NULL)
}
\arguments{
\item{key}{\code{string} The API key which is automatically loaded using the loadapikey() internal function.}

\item{quietly}{\code{logical}. To indicate if the token is successfully generated. Default \code{TRUE}.}

\item{seed}{\code{integer}. An integer to help track the caching of the access token generated during data collation.
If a user wants to get a new token, then the seed should be changed.}
}
\value{
\code{string} token authentication token key
}
\description{
The function and updates the authentication token that is automatically generated every after
six hours by the servers. The function gets the API key, which is a one time key that is provided
during registration or by database managers for already registered users. Since the authentication
token expires, the seed parameter is included to allow caching across user sessions. Therefore,
the data downloaded with a particular seed will be in stored in memory and can be retrieved from the user
PC than from servers and hence tremendously optimizing on the speed on data access.
}
\examples{

\dontrun{

#sessionkey <- secret_make_key()

#edit this page with usethis::edit_r_environ()

#enc_api <- secret_encrypt(x = 'apikey', key = sessionkey)

#encrypted token for my api key

enc_api <- "p6-9gAvYXveyC_B_0hTzyYl5FwLRwZEPD-ZE9Y4KrIBstgNc8K2Mr4u6t39LGZ2E1m9SOw"

#the FWTRAITS_KEY is the unlock key saved in my local environment
#check https://httr2.r-lib.org/articles/wrapping-apis.html for more information

#download fish catchment region data
#setting the FWTRAITS_KEY

#run this usethis::edit_r_environ()

apikeydecrypted <- loadapikey(test = TRUE, sacrambled_apikey = enc_api,
                              fwtraitskey =  'FWTRAITS_KEY')

tokendata <- fw_token(key= apikeydecrypted, seed = 1234)

}


}
