% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_vizualize.R
\name{fw_visualize}
\alias{fw_visualize}
\title{ggplot2 visualization.}
\usage{
fw_visualize(
  output,
  scales = "free",
  ncol = 2,
  params = NULL,
  group = NULL,
  viridis = TRUE
)
}
\arguments{
\item{output}{fetchdata output only accepted.}

\item{scales}{indicate if the scales are free, free_x, and free_y.}

\item{ncol}{Number of columns to display the data in facet_wrap.}

\item{params}{species ecological parameters selected if there are more than one.}

\item{group}{Particular organism group to filter out to allow visualization.}

\item{viridis}{\code{logical} Either to use scale_fill_viridis or scale_fill_grey}
}
\value{
ggplot2 display
}
\description{
ggplot2 visualization.
}
