% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Title}
\usage{
getdata(
  taxa,
  taxaorder = NULL,
  token,
  multiple = FALSE,
  parallel = FALSE,
  cores = NULL,
  seed = 1144
)
}
\arguments{
\item{cores}{}
}
\description{
Title
}
