% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invertbackbone}
\alias{invertbackbone}
\title{Taxonomic data for species in the freshwaterecology.info database}
\format{
A \code{tibble} 10421 rows and 3 columns.
}
\usage{
data(invertbackbone)
}
\description{
A \code{tibble}
}
\details{
Standard taxonomic backbone for macroinvertebrates from www.freshwaterecology.info
}
\examples{

\dontrun{

data("invertbackbone")
invertbackbone
}


}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). Www.freshwaterecology.info - An online tool that unifies,
standardises and codifies more than 20,000 European freshwater organisms and their ecological preferences.
Ecological Indicators, 53, 271-282. https://doi.org/10.1016/j.ecolind.2015.02.007.
}
\keyword{Macroinvertebrates}
