% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbenthodata}
\alias{pbenthodata}
\title{Taxonomic data for the phytobentho species in www.freshwaterecology.info database.}
\format{
A \code{tibble} The dataset has 1857 rows and 2 columns. The column names constains

\itemize{
\item Taxgroup, which is large taxonomic group naming containing multiple species.
\ item Taxon, is the species taxonomic names.
}
}
\usage{
data(pbenthodata)
}
\description{
A \code{tibble}
}
\details{
The phytobentho data forms the taxonomic backbone for the species archived in the database.
This enables taxonomic name checks and faster data retrieval from the database.
}
\examples{

\dontrun{

data("pbenthodata")

pbenthodata
}

}
\references{
Schmidt-Kloiber, A., & Hering, D. (2015). Www.freshwaterecology.info - An online tool that unifies,
standardises and codifies more than 20,000 European freshwater organisms and their ecological preferences.
Ecological Indicators, 53, 271-282. https://doi.org/10.1016/j.ecolind.2015.02.007.
}
\keyword{Phytobentho}
\keyword{data}
